% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_data.R
\name{biomes}
\alias{biomes}
\title{Load a spatial vector of eco-regions
Load a spatial vector of eco-regions as an sf object}
\usage{
biomes()
}
\value{
an sf object of eco-regions
}
\description{
Load a spatial vector of eco-regions
Load a spatial vector of eco-regions as an sf object
}
