% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{chml_set_options}
\alias{chml_set_options}
\title{Safely set chmloader options
Helper function to safely set chmloader options such as the output raster type}
\usage{
chml_set_options(out_raster_type = c("SpatRaster", "character"))
}
\arguments{
\item{out_raster_type}{character, the output raster type}
}
\description{
Safely set chmloader options
Helper function to safely set chmloader options such as the output raster type
}
\details{
This function can be used to set specific options for the chmloader
package. The only option currently available is \code{out_raster_type} which
determines the class of the output raster object from the \code{download_chm}
function. The default is "SpatRaster", but alternatively it can be set to
"character" to return the path to the downloaded CHM data.
}
