% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_chms.R
\name{compare_models}
\alias{compare_models}
\title{Standardised plot to compare two Canopy Height Models
Create a multi-panel plot showing the benchmark raster alongside the
meta/WRI raster, and a 2D density plot of the correspondence between the two
rasters.}
\usage{
compare_models(
  r1,
  r2,
  fill_trans = c("sqrt", "identity", "log", "log10"),
  title = "",
  aggregate = NULL,
  drop_zeros = FALSE,
  n_2d = 100
)
}
\arguments{
\item{r1}{a SpatRaster object, The benchmark raster (user provided)}

\item{r2}{a SpatRaster object, The meta/WRI raster. If this is not provided,
the function will download the missing raster for the area coincident with
\code{r1}.}

\item{fill_trans}{a character string of the transformation to apply to the
density colour gradient.}

\item{title}{a character string of the plot title}

\item{aggregate}{a numeric vector of the factors to aggregate the rasters by.
default is NULL in which case the CHMs are compared only at the native
resolution of \code{r1}. the aggregate factor is relative to the native resolution
of the rasters. see \code{terra::aggregate} for more details. Multiple numeric
values can be provided to compare the rasters at many different resolutions.}

\item{drop_zeros}{a logical, whether to drop zeros from the 2D density plot
dataframe and comparative statistics. If TRUE, when both the benchmark and
meta/WRI CHM are zero, these rows are removed. useful when there is a lot
of bare ground. default is TRUE}

\item{n_2d}{an integer, Number of grid points in each direction for the 2D
density. default is 100. see \code{ggplot::stat_density2d} for more details.}
}
\value{
a ggplot object if \code{aggregate} is NULL, otherwise a list of ggplot
objects.
}
\description{
Standardised plot to compare two Canopy Height Models
Create a multi-panel plot showing the benchmark raster alongside the
meta/WRI raster, and a 2D density plot of the correspondence between the two
rasters.
}
\details{
This function provides a standardised way to evaluate the accuracy
of two canopy height models (or indeed any two rasters with equal
dimensions).

The new Tolan, et al. (2024) global CHM presents a step change
in forest mapping. With this change comes some potential issues and varying
performance, but given how new these data are and the global extent, it
is challeging to develop a comprehensive evaluation. This
function offers a means by which to quickly and easily compare these data
with existing CHMs, derived either from LiDAR or through Machine Learning
workflows.

We hope that this function can help us collectively learn more about the
relative merits of this dataset and how we may use it and the underlying
models (see https://github.com/facebookresearch/HighResCanopyHeight) to
improve our understanding of forest structure and function.

The plot generates some general location information for the target area
including the continent and biome. Further information can be added to the
\code{title} argument if desired. We hope this reduces any barriers to sharing
the results.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

wist_wood <- reference_data("wistmans_wood")
compare_models(wist_wood, aggregate = 10, drop_zeros = TRUE)
\dontshow{\}) # examplesIf}
}
