% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_chms.R
\name{comparison_plot}
\alias{comparison_plot}
\title{Standardised plot to compare two Canopy Height Models
Create a multi-panel plot showing the benchmark raster alongside the
meta/WRI raster, and a 2D density plot of the correspondence between the two
rasters.}
\usage{
comparison_plot(
  r1,
  r2,
  fill_trans = c("sqrt", "identity", "log", "log10"),
  title = "",
  drop_zeros = TRUE
)
}
\arguments{
\item{r1}{a SpatRaster object, The benchmark raster (user provided)}

\item{r2}{a SpatRaster object, The meta/WRI raster}

\item{fill_trans}{a character string of the transformation to apply to the
density colour gradient.}

\item{title}{a character string of the plot title}

\item{drop_zeros}{a logical, whether to drop zeros from the 2D density plot
dataframe when both the benchmark and meta/WRI CHM are zero. default is TRUE}
}
\value{
a ggplot object
}
\description{
Standardised plot to compare two Canopy Height Models
Create a multi-panel plot showing the benchmark raster alongside the
meta/WRI raster, and a 2D density plot of the correspondence between the two
rasters.
}
