% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_data.R
\name{continents}
\alias{continents}
\title{Load a spatial vector of contentinents
Load a spatial vector of continents as an sf object}
\usage{
continents()
}
\value{
an sf object of continents
}
\description{
Load a spatial vector of contentinents
Load a spatial vector of continents as an sf object
}
