% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_data.R
\name{reference_data}
\alias{reference_data}
\title{Load example LiDAR-based CHM data.
function to access a set of example CHM data for demonstration and comparison
with the Tolan, et. al. (2024) CHM data.}
\usage{
reference_data(name = c("wistmans_wood", "fingle_woods", "new_forest"))
}
\arguments{
\item{name}{character, the name of the example CHM data to load}
}
\value{
a SpatRaster object
}
\description{
Load example LiDAR-based CHM data.
function to access a set of example CHM data for demonstration and comparison
with the Tolan, et. al. (2024) CHM data.
}
\details{
At present these examples include data from the Environment Agency's
Vegetation Object Model (VOM). See https://www.data.gov.uk/dataset/227ab487-e8f2-4cbb-b26a-9e6d3b662265/lidar-vegetation-object-model-vom
for more information. If you want to explore these data further in R, check
out the gblidar package at: https://github.com/h-a-graham/gblidar
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
wistmans_wood <- load_chm_data("wistmans_wood")
\dontshow{\}) # examplesIf}
}
